+++ enum_map/lib/annotations.dart
@@ -1,3 +1,3 @@
-export 'src/abstract_enum_map.dart';
-export 'src/enum_map.dart';
-export 'src/unmodifiable_enum_map.dart';
+export 'src/annotations/abstract_generate_enum_map.dart';
+export 'src/annotations/generate_enum_map.dart';
+export 'src/annotations/generate_unmodifiable_enum_map.dart';