+++ enum_map/lib/enum_map.dart
@@ -1,0 +1,4 @@
+export 'annotations.dart';
+
+export 'src/maps/enum_map.dart';
+export 'src/maps/unmodifiable_enum_map.dart';