+++ enum_map/lib/src/abstract_enum_map.dart
@@ -1,5 +1,0 @@
-/// A superclass for annotations that you actually use.
-abstract class AbstractEnumMap {
-  ///
-  const AbstractEnumMap();
-}