+++ enum_map/lib/src/annotations/abstract_generate_enum_map.dart
@@ -1,0 +1,5 @@
+/// A superclass for annotations that you actually use.
+abstract class AbstractGenerateEnumMap {
+  ///
+  const AbstractGenerateEnumMap();
+}