+++ enum_map/lib/src/annotations/generate_unmodifiable_enum_map.dart
@@ -1,12 +1,12 @@
-import 'abstract_enum_map.dart';
+import 'abstract_generate_enum_map.dart';
 
 /// Use this annotation on enums to create an unmodifiable map based on them.
 /// Such map has a const constructor.
-class UnmodifiableEnumMap extends AbstractEnumMap {
+class GenerateUnmodifiableEnumMap extends AbstractGenerateEnumMap {
   ///
-  const UnmodifiableEnumMap();
+  const GenerateUnmodifiableEnumMap();
 }
 
 /// Use this annotation on enums to create a modifiable map based on them.
 /// Such map has a const constructor.
-const unmodifiableEnumMap = UnmodifiableEnumMap();
+const unmodifiableEnumMap = GenerateUnmodifiableEnumMap();