+++ enum_map_gen/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.0
+
+* Require `enum_map` 0.2.0, see its changelog.
+
 ## 0.1.0
 
 * Initial release.